+++ chain/messagepool/messagepool.go
@@ -9,6 +9,7 @@ import (
 	stdbig "math/big"
 	"sort"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	"github.com/hashicorp/go-multierror"
@@ -114,6 +115,11 @@ type MessagePoolEvtMessage struct {
 	CID cid.Cid
 }
 
+type MessagePoolPendingSnapshot struct {
+	msgs []*types.SignedMessage
+	ts   *types.TipSet
+}
+
 func init() {
 	// if the republish interval is too short compared to the pubsub timecache, adjust it
 	minInterval := pubsub.TimeCacheDuration + time.Duration(buildconstants.PropagationDelaySecs)*time.Second
@@ -178,6 +184,8 @@ type MessagePool struct {
 
 	evtTypes [3]journal.EventType
 	journal  journal.Journal
+
+	snapshot atomic.Pointer[MessagePoolPendingSnapshot]
 }
 
 type stateNonceCacheKey struct {
@@ -1029,6 +1037,8 @@ func (mp *MessagePool) addLocked(ctx context.Context, m *types.SignedMessage, st
 	// Record the current size of the Mpool
 	stats.Record(ctx, metrics.MpoolMessageCount.M(int64(mp.currentSize)))
 
+	mp.snapshot.Store(nil)
+
 	return nil
 }
 
@@ -1224,16 +1234,32 @@ func (mp *MessagePool) remove(ctx context.Context, from address.Address, nonce u
 
 	// Record the current size of the Mpool
 	stats.Record(ctx, metrics.MpoolMessageCount.M(int64(mp.currentSize)))
+
+	mp.snapshot.Store(nil)
 }
 
 func (mp *MessagePool) Pending(ctx context.Context) ([]*types.SignedMessage, *types.TipSet) {
-	mp.curTsLk.RLock()
-	defer mp.curTsLk.RUnlock()
+	snapshot := mp.snapshot.Load()
 
-	mp.lk.RLock()
-	defer mp.lk.RUnlock()
+	if snapshot == nil {
+		mp.curTsLk.RLock()
+		defer mp.curTsLk.RUnlock()
 
-	return mp.allPending(ctx)
+		mp.lk.RLock()
+
+		msgs, ts := mp.allPending(ctx)
+		snapshot = &MessagePoolPendingSnapshot{
+			msgs: msgs,
+			ts:   ts,
+		}
+		mp.snapshot.Store(snapshot)
+
+		mp.lk.RUnlock()
+	}
+
+	msgs := make([]*types.SignedMessage, len(snapshot.msgs))
+	copy(msgs, snapshot.msgs)
+	return msgs, snapshot.ts
 }
 
 func (mp *MessagePool) allPending(ctx context.Context) ([]*types.SignedMessage, *types.TipSet) {
@@ -1612,6 +1638,8 @@ func (mp *MessagePool) loadLocal(ctx context.Context) error {
 func (mp *MessagePool) Clear(ctx context.Context, local bool) {
 	mp.lk.Lock()
 	defer mp.lk.Unlock()
+
+	mp.snapshot.Store(nil)
 
 	// remove everything if local is true, including removing local messages from
 	// the datastore